@@ -1,148 +1,157 @@
 {
-  "name": "react-semantic.ui-starter",
-  "version": "3.0.0",
-  "main": "common/index.jsx",
-  "description": "React/Redux/Jest/Webpack/Semantic.UI boilerplate/example.",
-  "dependencies": {
-    "classnames": "^2.2.5",
-    "cookie-parser": "^1.4.3",
-    "helmet": "^3.6.1",
-    "history": "^4.6.3",
-    "js-cookie": "^2.1.4",
-    "lodash": "^4.17.4",
-    "normalizr": "^3.2.3",
-    "offline-plugin": "^4.8.1",
-    "prop-types": "^15.5.10",
-    "rapscallion": "^2.1.12",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-helmet": "^5.1.3",
-    "react-redux": "^5.0.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "redux": "^3.7.1",
-    "redux-thunk": "^2.2.0",
-    "semantic-ui-css": "^2.2.10",
-    "semantic-ui-react": "^0.71.0",
-    "store2": "^2.5.1",
-    "styled-components": "^2.1.1",
-    "whatwg-fetch": "^2.0.3"
-  },
-  "devDependencies": {
-    "autoprefixer": "^7.1.2",
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-dynamic-import-node": "^1.0.2",
-    "babel-plugin-lodash": "^3.2.11",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-plugin-transform-react-jsx-img-import": "^0.1.4",
-    "babel-plugin-transform-regenerator": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-plugin-webpack-alias": "^2.1.2",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-optimize": "^1.0.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "6.24.1",
-    "babel-runtime": "^6.23.0",
-    "babili-webpack-plugin": "^0.1.2",
-    "compression-webpack-plugin": "^0.4.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.4",
-    "enzyme": "^2.9.1",
-    "esdoc": "^0.5.2",
-    "eslint": "^4.2.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-babel": "^4.1.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.1.0",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^7.1.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "^0.11.2",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "git-rev-sync": "^1.9.1",
-    "glob": "^7.1.2",
-    "html-webpack-plugin": "^2.29.0",
-    "husky": "^0.14.3",
-    "i18n-webpack-plugin": "^1.0.0",
-    "jest": "^20.0.4",
-    "json-server": "^0.11.0",
-    "minimist": "^1.2.0",
-    "node-fetch": "^1.7.1",
-    "node-sass": "^4.5.3",
-    "nodemon": "^1.11.0",
-    "open": "0.0.5",
-    "optimize-css-assets-webpack-plugin": "^2.0.0",
-    "postcss": "^6.0.6",
-    "postcss-browser-reporter": "^0.5.0",
-    "postcss-cssnext": "^3.0.0",
-    "postcss-import": "^10.0.0",
-    "postcss-loader": "^2.0.6",
-    "postcss-reporter": "^4.0.0",
-    "preload-webpack-plugin": "^1.2.2",
-    "purify-css": "^1.2.5",
-    "purifycss-webpack": "^0.7.0",
-    "react-addons-perf": "^15.4.2",
-    "react-addons-test-utils": "^15.6.0",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "redux-devtools-extension": "^2.13.2",
-    "redux-mock-store": "^1.2.3",
-    "resolve-url-loader": "^2.1.0",
-    "sass-loader": "^6.0.6",
-    "serve": "^6.0.2",
-    "style-loader": "^0.18.2",
-    "stylelint": "^7.12.0",
-    "stylelint-config-standard": "^16.0.0",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.1.0",
-    "webpack-bundle-analyzer": "^2.8.2",
-    "webpack-common-shake": "^1.5.0",
-    "webpack-dev-middleware": "^1.11.0",
-    "webpack-hot-middleware": "^2.18.2",
-    "webpack-manifest-plugin": "^1.1.2",
-    "webpack-subresource-integrity": "^1.0.1",
-    "why-did-you-update": "^0.0.8"
-  },
-  "optionalDependencies": {
-    "img-loader": "^2.0.0",
-    "prepack-webpack-plugin": "^1.1.0",
-    "uglify-js": "git://github.com/mishoo/UglifyJS2#harmony-v2.8.22",
-    "uglifyjs-webpack-plugin": "^0.4.6"
-  },
-  "scripts": {
-    "db": "nodemon ./db",
-    "docs": "esdoc && serve ./esdoc",
-    "serve_dist": "serve -s ./dist/$APP_LANGUAGE",
-    "build": "webpack --config webpack_config/webpack.prod.js",
-    "build:demo": "BUILD_DEMO=true APP_LANGUAGE=en webpack --config webpack_config/webpack.prod.demo.remove.me.please.js",
-    "test": "jest --config=jest_config/jest.config.json --coverage --forceExit || true",
-    "dev": "node webpack_config/server.js",
-    "start": "npm run dev",
-    "ssr": "nodemon --exec BABEL_ENV=ssr babel-node ./src/server",
-    "precommit": "npm run test",
-    "prepush": "echo \"$(tput bold)$(tput setaf 5)Hello, my name is React-Semantic-UI-Starter, and I am happy to see that you use this project :) $(tput sgr 0)\"",
-    "lint:styles": "stylelint ./src/common/**/*.scss -s scss --config ./.stylelintrc --fix",
-    "lint:scripts": "eslint -c ./.eslintrc.json ./src ./webpack_config ./db --ignore-pattern dist"
-  },
-  "author": "Vladimir Metnew <vladimirmetnew@gmail.com>",
-  "license": "MIT",
-  "repository": "https://github.com/Metnew/react-semantic.ui-starter.git",
-  "keywords": [
-    "react",
-    "redux",
-    "react-redux",
-    "starter",
-    "template",
-    "boilerplate",
-    "webpack"
-  ]
+	"name": "react-semantic.ui-starter",
+	"version": "3.0.0",
+	"main": "src/client/index.jsx",
+	"description": "React/Redux/Jest/Webpack/Semantic.UI boilerplate/example.",
+	"dependencies": {
+		"babel-core": "^6.25.0",
+		"body-parser": "^1.17.2",
+		"chalk": "^2.0.1",
+		"classnames": "^2.2.5",
+		"compression": "^1.7.0",
+		"cookie-parser": "^1.4.3",
+		"es6-promise": "^4.1.1",
+		"express": "^4.15.3",
+		"helmet": "^3.8.1",
+		"history": "^4.6.3",
+		"isomorphic-fetch": "^2.2.1",
+		"js-cookie": "^2.1.4",
+		"jsonwebtoken": "^7.4.2",
+		"lodash": "^4.17.4",
+		"morgan": "^1.8.2",
+		"normalizr": "^3.2.3",
+		"offline-plugin": "^4.8.3",
+		"prop-types": "^15.5.10",
+		"rapscallion": "^2.1.13",
+		"react": "^15.6.1",
+		"react-dom": "^15.6.1",
+		"react-helmet": "^5.1.3",
+		"react-redux": "^5.0.5",
+		"react-router": "^4.1.2",
+		"react-router-dom": "^4.1.2",
+		"react-router-redux": "^5.0.0-alpha.6",
+		"redux": "^3.7.2",
+		"redux-thunk": "^2.2.0",
+		"semantic-ui-css": "^2.2.11",
+		"semantic-ui-react": "^0.71.3",
+		"store2": "^2.5.1",
+		"styled-components": "^2.1.2"
+	},
+	"devDependencies": {
+		"autoprefixer": "^7.1.2",
+		"babel-cli": "^6.24.1",
+		"babel-eslint": "^7.2.3",
+		"babel-loader": "^7.1.1",
+		"babel-plugin-dynamic-import-node": "^1.0.2",
+		"babel-plugin-lodash": "^3.2.11",
+		"babel-plugin-syntax-dynamic-import": "^6.18.0",
+		"babel-plugin-transform-react-jsx": "^6.24.1",
+		"babel-plugin-transform-regenerator": "^6.24.1",
+		"babel-plugin-transform-runtime": "^6.23.0",
+		"babel-plugin-webpack-alias": "^2.1.2",
+		"babel-polyfill": "^6.23.0",
+		"babel-preset-es2015": "^6.24.1",
+		"babel-preset-react": "^6.24.1",
+		"babel-preset-react-optimize": "^1.0.1",
+		"babel-preset-stage-0": "^6.24.1",
+		"babel-register": "6.24.1",
+		"babel-runtime": "^6.25.0",
+		"babili-webpack-plugin": "^0.1.2",
+		"compression-webpack-plugin": "^1.0.0",
+		"concurrently": "^3.5.0",
+		"css-loader": "^0.28.4",
+		"esdoc": "^1.0.1",
+		"eslint": "^4.4.0",
+		"eslint-config-standard": "^10.2.1",
+		"eslint-loader": "^1.9.0",
+		"eslint-plugin-babel": "^4.1.2",
+		"eslint-plugin-import": "^2.7.0",
+		"eslint-plugin-node": "^5.1.1",
+		"eslint-plugin-promise": "^3.5.0",
+		"eslint-plugin-react": "^7.1.0",
+		"eslint-plugin-standard": "^3.0.1",
+		"extract-text-webpack-plugin": "^3.0.0",
+		"favicons-webpack-plugin": "0.0.7",
+		"file-loader": "^0.11.2",
+		"friendly-errors-webpack-plugin": "^1.6.1",
+		"git-rev-sync": "^1.9.1",
+		"glob": "^7.1.2",
+		"html-webpack-plugin": "^2.30.1",
+		"http-proxy": "^1.16.2",
+		"husky": "^0.14.3",
+		"i18n-webpack-plugin": "^1.0.0",
+		"img-loader": "^2.0.0",
+		"jest": "^20.0.4",
+		"minimist": "^1.2.0",
+		"node-noop": "^1.0.0",
+		"node-sass": "^4.5.3",
+		"nodemon": "^1.11.0",
+		"open": "0.0.5",
+		"optimize-css-assets-webpack-plugin": "^3.0.0",
+		"postcss": "^6.0.8",
+		"postcss-browser-reporter": "^0.5.0",
+		"postcss-cssnext": "^3.0.2",
+		"postcss-import": "^10.0.0",
+		"postcss-loader": "^2.0.6",
+		"postcss-reporter": "^5.0.0",
+		"preload-webpack-plugin": "^1.2.3",
+		"purify-css": "^1.2.5",
+		"purifycss-webpack": "^0.7.0",
+		"react-addons-perf": "^15.4.2",
+		"react-addons-test-utils": "^15.6.0",
+		"react-hot-loader": "^3.0.0-beta.6",
+		"redux-devtools-extension": "^2.13.2",
+		"redux-mock-store": "^1.2.3",
+		"resolve-url-loader": "^2.1.0",
+		"sass-loader": "^6.0.6",
+		"serve": "^6.0.5",
+		"style-loader": "^0.18.2",
+		"stylelint": "^8.0.0",
+		"stylelint-config-standard": "^17.0.0",
+		"uglify-js": "git://github.com/mishoo/UglifyJS2#harmony-v2.8.22",
+		"uglifyjs-webpack-plugin": "^0.4.6",
+		"url-loader": "^0.5.9",
+		"webpack": "^3.4.1",
+		"webpack-bundle-analyzer": "^2.9.0",
+		"webpack-common-shake": "^1.5.3",
+		"webpack-dev-middleware": "^1.12.0",
+		"webpack-hot-middleware": "^2.18.2",
+		"webpack-manifest-plugin": "^1.2.1",
+		"webpack-shell-plugin": "^0.5.0",
+		"webpack-subresource-integrity": "^1.0.1",
+		"why-did-you-update": "^0.0.8"
+	},
+	"optionalDependencies": {
+		"prepack-webpack-plugin": "^1.1.0"
+	},
+	"scripts": {
+		"docs": "esdoc && serve ./esdoc",
+		"build": "npm run frontend_build && npm run server_build",
+		"dev": "concurrently -p \"[{name}]\" -n  \"SERVER,FRONTEND\" -c \"cyan.bold,magenta.bold\" \"npm run server_dev\" \"npm run frontend_dev\"",
+		"server_dev": "webpack --config webpack_config/server/webpack.dev.babel.js",
+		"server_build": "NODE_ENV=production webpack --config webpack_config/server/webpack.prod.babel.js",
+		"test": "jest --config=jest_config/jest.config.json --coverage --forceExit || true",
+		"frontend_dev": "babel-node webpack_config/server.js",
+		"frontend_build": "NODE_ENV=production webpack --config webpack_config/client/webpack.prod.babel.js",
+		"start": "node ./dist/server",
+		"precommit": "echo \"$(tput bold)$(tput setaf 5) It'd be better to run tests before commit. $(tput sgr 0)\"t",
+		"prepush": "echo \"$(tput bold)$(tput setaf 5) I hope you tested app. $(tput sgr 0)\"",
+		"lint:styles": "stylelint ./src/common/**/*.scss -s scss --config ./.stylelintrc --fix",
+		"lint:scripts": "eslint -c ./.eslintrc.json ./src ./webpack_config --ignore-pattern dist"
+	},
+	"engines": {
+		"node": ">=6"
+	},
+	"author": "Vladimir Metnew <vladimirmetnew@gmail.com>",
+	"license": "MIT",
+	"repository": "https://github.com/Metnew/react-semantic.ui-starter.git",
+	"keywords": [
+		"react",
+		"redux",
+		"react-redux",
+		"starter",
+		"template",
+		"boilerplate",
+		"webpack"
+	]
 }
