@@ -1,2 +1,2 @@
-global.fetch = require('node-fetch')
+global.fetch = require('isomorphic-fetch')
 process.env.BASE_API = process.env.BASE_API || 'http://localhost:4000/api/v1'
