@@ -1,16 +1,26 @@
 import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import {Provider} from 'react-redux'
-import {ConnectedRouter as Router} from 'react-router-redux'
 import {APPLICATION_INIT} from 'actions'
 import {ThemeProvider} from 'styled-components'
 import theme from 'styles/theme'
+import App from 'containers/App'
+import RoutingWrapper from 'components/addons/RoutingWrapper'
+
+const Router = process.env.BROWSER === true
+  ? require('react-router-redux').ConnectedRouter
+  : require('react-router').StaticRouter
 
 export default class Root extends Component {
   static propTypes = {
     store: PropTypes.object,
+    SSR: PropTypes.object,
     history: PropTypes.object,
-    routes: PropTypes.func
+    routes: PropTypes.array
+  }
+
+  static defaultProps = {
+    SSR: {}
   }
 
   componentWillMount () {
@@ -37,13 +47,18 @@ export default class Root extends Component {
   }
 
   render () {
-    const {store, history, routes} = this.props
+    const {SSR, store, history, routes} = this.props
+    const routerProps = process.env.BROWSER === true
+      ? {history}
+      : {location: SSR.location, context: SSR.context}
     // key={Math.random()} = hack for HMR from https://github.com/webpack/webpack-dev-server/issues/395
     return (
       <Provider store={store} key={Math.random()}>
-        <Router history={history} key={Math.random()}>
+        <Router {...routerProps} key={Math.random()}>
           <ThemeProvider theme={theme}>
-            {routes(::this.authCheck)}
+            <App>
+              <RoutingWrapper routes={routes} authCheck={::this.authCheck}/>
+            </App>
           </ThemeProvider>
         </Router>
       </Provider>
