@@ -12,7 +12,7 @@ class UsersItemComponent extends Component {
     const {user} = this.props
     return (
       <Card>
-        <Image alt={`${user.name}`} src="./images/daniel.jpg" />
+        <Image alt={`${user.name}`} src={require('images/daniel.jpg')} />
         <Card.Content>
           <Card.Header>
             {user.name}
