@@ -3,10 +3,11 @@ process.env.NODE_ENV = 'development'
 const path = require('path')
 const webpack = require('webpack')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
-const base = require('./webpack.base')
+const baseWebpackConfig = require('./webpack.base')
 const FriendlyErrors = require('friendly-errors-webpack-plugin')
+const ManifestPlugin = require('webpack-manifest-plugin')
+const config = require('../config')
 
-const config = require('./config')
 const loaders = {
   style: {loader: 'style-loader'},
   css: {loader: 'css-loader', options: {sourceMap: true}},
@@ -20,8 +21,8 @@ const loaders = {
   sass: {loader: 'sass-loader', options: {sourceMap: true}}
 }
 
-base.devtool = 'eval-source-map'
-base.module.rules.push(
+baseWebpackConfig.devtool = 'eval-source-map'
+baseWebpackConfig.module.rules.push(
   {
     test: /\.css$/,
     loaders: [loaders.style, loaders.css, loaders.postcss, loaders.resolve]
@@ -39,16 +40,17 @@ base.module.rules.push(
 )
 
 // add dev plugins
-base.plugins.push(
+baseWebpackConfig.plugins.push(
   // add index.html
   new HtmlWebpackPlugin({
     title: config.title,
-    template: path.resolve(config.srcCommonPath, 'index.html'),
-    filename: path.resolve(config.distPath, 'index.html')
+    template: path.resolve(config.srcCommonPath, 'index.ejs')
   }),
+  // create manifest.json
+  new ManifestPlugin({fileName: 'manifest.json', cache: config.manifest}),
   new webpack.HotModuleReplacementPlugin(),
   new webpack.NoEmitOnErrorsPlugin(),
   new FriendlyErrors()
 )
 
-module.exports = base
+module.exports = baseWebpackConfig
